# src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  title: API аутентификации
  description: Документация для регистрации и входа
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /api/cards/admin/get-all:
    get:
      tags:
        - Карты (Админ)
      summary: Получить все карты (с фильтрацией)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: userId
          in: query
          description: ID пользователя для фильтрации
          required: false
          schema:
            type: integer
            minimum: 1
        - name: status
          in: query
          description: Статус карты для фильтрации
          required: false
          schema:
            $ref: '#/components/schemas/BankCardStatus'
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankCardResponseDto'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/admin/create/{userId}:
    post:
      tags:
        - Карты (Админ)
      summary: Создать новую карту для пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '201':
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardResponseDto'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/admin/{cardId}/block:
    patch:
      tags:
        - Карты (Админ)
      summary: Заблокировать карту
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Карта заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardResponseDto'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/admin/{cardId}/activate:
    patch:
      tags:
        - Карты (Админ)
      summary: Активировать карту
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Карта активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardResponseDto'
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/admin/{cardId}:
    delete:
      tags:
        - Карты (Админ)
      summary: Удалить карту
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Карта удалена
        '400':
          description: Неверные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BankCardResponseDto:
      type: object
      properties:
        id:
          type: integer
          description: ID карты
          example: 1
        maskedCardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 1234"
        userId:
          type: integer
          description: ID владельца карты
          example: 1
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2025-12-31"
        status:
          $ref: '#/components/schemas/BankCardStatus'
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.50
...
    BankCardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED
      description: Статус банковской карты
      example: ACTIVE
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Код ошибки
          example: VALIDATION_ERROR
        errorMessage:
          type: string
          description: Сообщение об ошибке
          example: Email should be valid
security:
  - bearerAuth: []